requirements.txt
torch
numpy
Pillow
tabulate
tqdm
torchvision

gdown